refactor: bump python function app to 3.9
