+++ .github/workflows/python.yml
@@ -8,7 +8,7 @@ on:
 env:
   AZURE_FUNCTIONAPP_NAME: freshfarmfunctions
   AZURE_FUNCTIONAPP_PACKAGE_PATH: "backend/python"
-  PYTHON_VERSION: "3.7"
+  PYTHON_VERSION: "3.9"
 
 jobs:
   build-and-deploy:
