+++ .gitignore
@@ -1,1 +1,1 @@
-.vscode/settings.json
+.vscode
