--- backend/python/python-test-2/sample.dat
@@ -1,3 +1,0 @@
-{
-    "name": "Azure"
-}
