--- backend/python/python-test/function.json
@@ -1,20 +1,0 @@
-{
-	"scriptFile": "__init__.py",
-	"bindings": [
-		{
-			"authLevel": "function",
-			"type": "httpTrigger",
-			"direction": "in",
-			"name": "req",
-			"methods": [
-				"get",
-				"post"
-			]
-		},
-		{
-			"type": "http",
-			"direction": "out",
-			"name": "$return"
-		}
-	]
-}
